backend/requirements.txt
africastalking==1.2.9
aiodns==3.5.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.17.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.2
asttokens==3.0.0
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.1.0
audioop-lts==0.2.2
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
babel==2.17.0
bcolz-zipline==1.13.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
blosc2==3.10.2
Bottleneck==1.6.0
CacheControl==0.14.3
cachetools==6.2.1
cctx==1.0.0
ccxt==4.5.3
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
conformer==0.3.2
constantly==23.10.4
contourpy==1.3.3
cron-descriptor==1.4.5
cryptography==44.0.2
curl_cffi==0.13.0
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-admin-black==1.0.15
django-allauth==65.4.1
django-appconf==1.1.0
django-browser-reload==1.18.0
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
django-pwa==2.0.1
django-select2==8.4.0
django-sendsms==0.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.0
einops==0.8.1
empyrical-reloaded==0.5.12
exchange_calendars==4.11.2
executing==2.2.1
fastapi==0.115.12
fastjsonschema==2.21.2
filelock==3.19.1
firebase_admin==7.1.0
flatbuffers==25.2.10
fonttools==4.60.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.9.0
gdown==5.2.0
geographiclib==2.0
geoip2==5.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.27.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.2
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.14.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
hyperlink==21.0.0
HyperPyYAML==1.2.2
idna==3.10
incremental==24.7.2
inflect==7.5.0
inquirerpy==0.3.4
intervaltree==3.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
iso3166==2.1.1
iso4217==1.14.20250512
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kagglehub==0.3.13
kaldifst==1.7.17
kiwisolver==1.4.9
kombu==5.4.2
korean-lunar-calendar==0.3.1
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.0
lru-dict==1.3.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
maxminddb==2.6.3
mdurl==0.1.2
mistune==3.1.4
modelscope==1.30.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multipledispatch==1.0.0
multitasking==0.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numba==0.62.0
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.72.0
openai-whisper==20250625
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
paypalrestsdk==1.13.3
peewee==3.17.3
pexpect==4.9.0
pfzy==0.3.4
phonenumbers==9.0.2
pillow==11.1.0
platformdirs==4.3.8
pooch==1.8.2
postgis==1.0.4
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.10.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.2
PyJWT==2.10.1
pyluach==2.3.0
pyOpenSSL==25.0.0
pyparsing==3.2.5
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.3.0
python3-openid==3.2.0
pytorch-lightning==2.5.5
pytz==2025.2
pyworld==0.3.5
PyYAML==6.0.2
pyzmq==27.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
schema==0.7.7
scikit-learn==1.7.1
scipy==1.16.1
segno==1.6.1
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.46.2
statsmodels==0.14.5
sympy==1.14.0
ta==0.11.0
tables==3.10.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
toolz==1.1.0
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
twilio==9.4.4
Twisted==24.11.0
txaio==23.1.1
typeguard==4.4.4
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wetext==0.1.0
wget==3.2
wheel==0.45.1
whitenoise==6.9.0
xgboost==3.0.4
yarl==1.18.3
yfinance==0.2.65
zipline-reloaded==3.1.1
zope.interface==7.2
